add project type handling in cleanappsettings (#1170)

+++ projectroles/management/commands/cleanappsettings.py
@@ -12,6 +12,10 @@ logger = ManagementCommandLogger(__name__)
 # Local constants
 START_MSG = 'Checking database for undefined app settings..'
 END_MSG = 'OK'
+DEFINITION_NOT_FOUND_MSG = 'definition not found'
+ALLOWED_TYPES_MSG = 'does not match allowed types'
+DELETE_PREFIX_MSG = 'Deleting "{}" from project "{}": '
+DELETE_PROJECT_TYPE_MSG = 'project type "{}" {}: {}'
 
 
 def get_setting_str(db_setting):
@@ -42,11 +46,27 @@ class Command(BaseCommand):
             else:
                 def_kwargs['app_name'] = 'projectroles'
             try:
-                app_settings.get_definition(**def_kwargs)
+                definition = app_settings.get_definition(**def_kwargs)
             except ValueError:
                 logger.info(
-                    'Deleting "{}" from project "{}"'.format(
+                    DELETE_PREFIX_MSG.format(
                         get_setting_str(s), s.project.title
+                    )
+                    + DEFINITION_NOT_FOUND_MSG
+                )
+                s.delete()
+                continue
+            if s.project and s.project.type not in definition.get(
+                'project_types', ['PROJECT']
+            ):
+                logger.info(
+                    DELETE_PREFIX_MSG.format(
+                        get_setting_str(s), s.project.title
+                    )
+                    + DELETE_PROJECT_TYPE_MSG.format(
+                        s.project.type,
+                        ALLOWED_TYPES_MSG,
+                        definition.get('project_types', ['PROJECT']),
                     )
                 )
                 s.delete()
